pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "weatherchannel"
version = "0.1.0"
description = "Deterministic Polymarket weather-market trading engine"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.5,<3.0",
    "pyyaml>=6.0,<7.0",
    "scipy>=1.12,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "respx>=0.21,<1.0",
    "ruff>=0.3,<1.0",
    "mypy>=1.8,<2.0",
]

[tool.setuptools.packages.find]
include = ["engine*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["engine/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
